NAT-301: use geographic columns for location for correct distance ordering
Currently, the location column is set as a non-geographic column, so when ordering by distance, the
database does not know how to deal with the fact the earth is spherical, so ordering can be wrong.
This tells the database that the column is geographic so the correct maths is applied when doing
geographic logic.