NAT-547: fix bug serialising offices with no location for EpiServer
EpiServer requires a latLong on every office, but this is optional in LSS.
It appears Resource Directory handled this by defaulting to co-ords of 0,0
so we replicate the same functionality. It doesn't really matter anyway as
volunteer search doesn't render a map so it's not used, but it's a case we
need to handle in EpiServer's parsing logic.