NAT-888: add many-to-many relationship between LCAs and local authorities
LSS now supports many-to-many relationships between LCAs and local authorities, as some LCAs serve
multiple council areas. At the moment we currently tell users there is no LCA in their area if they
are not in the primary local authority of that LCA, which is a poor user experience.

Network Data will have to update their data feed to us to expose this data, but in advance of that
we can change the relationships between local authorities and local offices to be a many-to-many
relationship. This table will be populated on import, and once the new table is deployed to prod
and the first import job has run, we can then remove the old belongs_to field on Office and start
querying using the new table. Until then, this leaves the old field and old query logic in place
so there is no disruption to the end user.