chore: maintenance to allow spinning up a new instance from scratch
Some of the migrations used syntax that had been obsoleted from Rails so no longer would apply cleanly
on a new database. Also the version of Traefik being used in the Docker compose was ancient and no
longer supported modern Docker, but it doesn't actually serve any purpose so we can get rid of it.